+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `center_of_mass` property to Assembly class.
+* Added `volume` property to Assembly class.
+
 ### Changed
 
 * Reduced some boilerplate code in `Joint` subclasses.